+++ lib/manageiq/providers/ibm_power_vc.rb
