+++ spec/spec_helper.rb
@@ -6,7 +6,7 @@ end
 Dir[Rails.root.join("spec/shared/**/*.rb")].each { |f| require f }
 Dir[File.join(__dir__, "support/**/*.rb")].each { |f| require f }
 
-require "manageiq-providers-ibm_power_vc"
+require "manageiq/providers/ibm_power_vc"
 
 def sanitizer(interaction)
   # Remove internal network hostnames
